% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ipadic.R
\name{check_ipadic}
\alias{check_ipadic}
\title{Check if IPADIC is properly installed}
\usage{
check_ipadic(
  dir = getOption("jliwc_project_home"),
  ipadic_dictname = getOption("jliwc_IPADIC_dir"),
  silent = FALSE
)
}
\arguments{
\item{dir}{A directory where IPADIC is installed}

\item{ipadic_dictname}{a name of the directory where IPADIC is installed}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful, \code{FALSE} otherwise
}
\description{
Check if IPADIC is properly installed
}
\examples{
\dontrun{
check_ipadic()
}
}
