% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mecab.R
\name{count_mecab}
\alias{count_mecab}
\title{Word segmentation using MeCab}
\usage{
count_mecab(
  text,
  text_field = "text",
  sys_dic = getOption("jliwc_IPADIC"),
  user_dic = getOption("jliwc_USERDIC"),
  liwclike = TRUE
)
}
\arguments{
\item{text}{A character vector}

\item{text_field}{A character vector}

\item{sys_dic}{A character vector}

\item{user_dic}{A character vector}

\item{liwclike}{A logical value}
}
\value{
A data frame
}
\description{
Word segmentation using MeCab
}
\examples{
setup_ipadic()
setup_userdic()
setup_jliwcdic()

gibasa::ginga[1:10] |> count_mecab()

}
