% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_info2.R
\name{dictionary_info2}
\alias{dictionary_info2}
\title{function that calls \code{\link[gibasa]{dictionary_info}} with locally installed MeCab}
\usage{
dictionary_info2(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[gibasa]{dictionary_info}}}
}
\value{
tokenized character vector
}
\description{
function that calls \code{\link[gibasa]{dictionary_info}} with locally installed MeCab
}
