% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dictionaries.R
\name{install_dictionaries}
\alias{install_dictionaries}
\title{Install dictionaries}
\usage{
install_dictionaries(ipadic = TRUE, userdic = TRUE, jliwcdic = TRUE)
}
\arguments{
\item{ipadic}{Install IPAdic (default: TRUE)}

\item{userdic}{Install the user dictionary (default: TRUE)}

\item{jliwcdic}{Install the J-LIWC2015 dictionary (default: TRUE)}
}
\description{
Install IPAdic, the user dictionary, and the J-LIWC2015 dictionaries.
}
\examples{
\dontrun{
install_dictionaries()
}

}
