% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ipadic.R
\name{install_ipadic}
\alias{install_ipadic}
\title{Download and install IPADIC}
\usage{
install_ipadic(
  dir = getOption("jliwc_project_home"),
  ipadic_dictname = getOption("jliwc_IPADIC_dir"),
  tar_dic = "mecab-ipadic-2.7.0-20070801",
  ipadic_url = getOption("jliwc_IPADIC_url"),
  silent = FALSE
)
}
\arguments{
\item{dir}{A directory where IPADIC is installed
(e.g., "C:/Users/username/Documents/J-LIWC2015/")}

\item{ipadic_dictname}{a name of the directory where IPADIC is installed
(e.g., "C:/Users/username/Documents/J-LIWC2015/mecab-ipadic")}

\item{tar_dic}{A name of the IPADIC tar file}

\item{ipadic_url}{A URL of the IPADIC tar file. Default on SourceForge.
If you fail to download the file from the default URL,
you can specify the URL of the tar file. For example, you can use the following URL:
\url{https://sourceforge.net/projects/mecab/files/mecab-ipadic/2.7.0-20070801/mecab-ipadic-2.7.0-20070801.tar.gz/download}}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful,
\code{FALSE} otherwise
}
\description{
Download and install IPADIC
}
\examples{
\dontrun{
install_ipadic()
}

}
