% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_jliwcdic.R
\name{install_jliwcdic}
\alias{install_jliwcdic}
\title{Install J-LIWC dictionary file}
\usage{
install_jliwcdic(
  dir = getOption("jliwc_project_home"),
  format = getOption("jliwc_format", default = "LIWC2015"),
  silent = FALSE
)
}
\arguments{
\item{dir}{a path to the J-LIWC2015 dictionary}

\item{format}{A format of the dictionary (LIWC2015 or LIWC22).
Use this option if the dictionary type is not automatically detected.}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful,
\code{FALSE} otherwise
}
\description{
Install J-LIWC dictionary file
}
\examples{
\dontrun{
install_jliwcdic()
}

}
