% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_userdic.R
\name{install_userdic}
\alias{install_userdic}
\title{Download and install J-LIWC2015 user dictionary}
\usage{
install_userdic(
  dir = getOption("jliwc_project_home"),
  user_dic = "user_dict.dic",
  userdic_url = getOption("jliwc_USERDIC_url"),
  silent = FALSE
)
}
\arguments{
\item{dir}{A home directory for J-LIWC2015}

\item{user_dic}{A file name of the user dictionary file}

\item{userdic_url}{A URL to the user dictionary file (default on GitHub)}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful,
\code{FALSE} otherwise
}
\description{
Download and install J-LIWC2015 user dictionary
}
\examples{
\dontrun{
install_userdic()
}

}
