% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{jliwc}
\alias{jliwc}
\alias{_PACKAGE}
\title{jliwc}
\description{
This package provides functions to analyze Japanese text using J-LIWC,
the Japanese version of LIWC (Linguistic Inquiry and Word Count).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tasukuigarashi/jliwc}
  \item Report bugs at \url{https://github.com/tasukuigarashi/jliwc/issues}
}

}
\author{
\strong{Maintainer}: Tasuku Igarashi \email{igarashi.tasuku@gmail.com} (\href{https://orcid.org/0000-0001-9432-4425}{ORCID})

}
