% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liwc_analysis.R
\name{liwc_analysis}
\alias{liwc_analysis}
\title{LIWC-style analysis of Japanese text data}
\usage{
liwc_analysis(
  input,
  dict,
  text_field = "text",
  lang = c("en", "ja"),
  pos_tag = TRUE,
  sys_dic = getOption("jliwc_IPADIC"),
  user_dic = getOption("jliwc_USERDIC")
)
}
\arguments{
\item{input}{A data frame or a list of character vector to be analyzed}

\item{dict}{A dictionary2 class object of J-LIWC2015 dictionary}

\item{text_field}{A column name of the input data frame to be analyzed}

\item{lang}{A character vector specifying language of the output tag (English or Japanese)}

\item{pos_tag}{logical; if TRUE, add POS percentage in MeCab}

\item{sys_dic}{A path to the MeCab system dictionary}

\item{user_dic}{A path to the MeCab user dictionary}
}
\value{
A data frame with LIWC-style analysis results
}
\description{
LIWC-style analysis of Japanese text data
}
\examples{
setup_ipadic()
setup_userdic()
setup_jliwcdic()

dict <- read_dict()

x <- gibasa::ginga[1:10] |> liwc_analysis(dict = dict)
x

}
