% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dictionaries.R
\name{load_dictionaries}
\alias{load_dictionaries}
\title{Load dictionaries}
\usage{
load_dictionaries(silent = FALSE)
}
\arguments{
\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the dictionaries are loaded successfully, \code{FALSE} otherwise
}
\description{
Load IPAdic, user dictionary, and J-LIWC dictionary.
Prior to using this function, you need to install the dictionaries by
\code{install_ipadic()}, \code{install_userdic()}, and \code{install_jliwcdic()}. You also need to set
\code{options(jliwc_project_home = "/path/to/directory")} where the dictionaries are installed.
}
\examples{
\dontrun{
options(jliwc_project_home = "/path/to/directory")

load_dictionaries()
}

}
