% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ipadic.R
\name{load_ipadic}
\alias{load_ipadic}
\title{Load IPADIC}
\usage{
load_ipadic(
  dir = getOption("jliwc_project_home"),
  ipadic_dictname = getOption("jliwc_IPADIC_dir"),
  silent = FALSE
)
}
\arguments{
\item{dir}{A directory where IPADIC is installed}

\item{ipadic_dictname}{a name of the directory where IPADIC is installed}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful, \code{FALSE} otherwise
}
\description{
Load IPADIC
}
\examples{
\dontrun{
load_ipadic()
}
}
