% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_jliwcdic.R
\name{load_jliwcdic}
\alias{load_jliwcdic}
\title{Load J-LIWC Dictionary File}
\usage{
load_jliwcdic(dir = getOption("jliwc_project_home"), silent = FALSE)
}
\arguments{
\item{dir}{The directory from which to load the J-LIWC dictionary file.
Defaults to the value set in the 'jliwc_project_home' option.
Note that it is not a path to a dictionary file, but a path to a
directory that contains the dictionary file.}

\item{silent}{Logical; if \code{FALSE}, prints messages to the console about the
loading process. Defaults to \code{FALSE}.}
}
\value{
Invisibly returns \code{TRUE} if the dictionary file is successfully loaded.
If an error occurs during the loading process, the function will stop and print
an error message. If a warning is raised during the loading process, it will
return \code{FALSE} and print the warning message.
}
\description{
This function loads the J-LIWC dictionary file from a specified directory.
It sets a temporary directory to avoid errors related to installing IPADIC
in a path that includes full-byte characters.
}
\examples{
\dontrun{
# To load the dictionary from the default project home directory:
load_jliwcdic()

# To load the dictionary from a specified directory and suppress messages:
load_jliwcdic("/path/to/dictionary", silent = TRUE)
}

}
