% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_userdic.R
\name{load_userdic}
\alias{load_userdic}
\title{Load user dictionary}
\usage{
load_userdic(
  dir = getOption("jliwc_project_home"),
  user_dic = "user_dict.dic",
  silent = FALSE
)
}
\arguments{
\item{dir}{A directory where IPADIC is installed}

\item{user_dic}{a name of the directory where a user dictionary is installed}

\item{silent}{Boolean. Whether to print messages}
}
\value{
Boolean, \code{TRUE} if the setup is successful, \code{FALSE} otherwise
}
\description{
Load user dictionary
}
\examples{
\dontrun{
load_userdic()
}
}
