% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing text for morphological analysis}
\usage{
preprocess(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A character vector. The text is converted to lowercase,
line breaks are replaced with \verb{\\n},
numbers are replaced with half-width characters,
and characters such as \code{!} and \code{@} are replaced with full-width characters.
}
\description{
Preprocessing text for morphological analysis
}
\examples{
\dontrun{
x <- preprocess("OK、今日の２時に！\r\n")
x
}

}
