% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_liwc.R
\name{preprocess_liwc}
\alias{preprocess_liwc}
\title{Morphological analysis imitating LIWC's internal processing}
\usage{
preprocess_liwc(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A character vector. Japanese characters that can be interpreted as
a word boundary \verb{\\b} in regular expressions are omitted from the original text.
}
\description{
Morphological analysis imitating LIWC's internal processing
}
