% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dict.R
\name{read_dict22}
\alias{read_dict22}
\title{Reading the dictionary and getting English categories (LIWC22 format)}
\usage{
read_dict22(file)
}
\arguments{
\item{file}{A file path of the dictionary}
}
\description{
Reading the dictionary and getting English categories (LIWC22 format)
}
