% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_setting_jliwc.R
\name{save_jliwc_config}
\alias{save_jliwc_config}
\title{Save and load a setting file}
\usage{
save_jliwc_config(file_path, config_key)
}
\arguments{
\item{file_path}{A character string of the file path}

\item{config_key}{A character string of the key to save the file path}
}
\description{
Load and save a setting file for jliwc package
}
