% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ipadic.R
\name{setup_ipadic}
\alias{setup_ipadic}
\title{Download and install IPADIC}
\usage{
setup_ipadic(
  dir = getOption("jliwc_project_home"),
  ipadic_dictname = "mecab-ipadic",
  tar_dic = "mecab-ipadic-2.7.0-20070801",
  ipadic_url = getOption("jliwc_IPADIC_url"),
  check = TRUE
)
}
\arguments{
\item{dir}{the directory where IPADIC is installed}

\item{ipadic_dictname}{the name of the directory where IPADIC is installed}

\item{tar_dic}{the name of the tar file}

\item{ipadic_url}{the URL of the tar file}

\item{check}{whether to check if IPADIC is installed}
}
\description{
Download and install IPADIC
}
\examples{
setup_ipadic()

}
