% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_jliwcdic.R
\name{setup_jliwcdic}
\alias{setup_jliwcdic}
\title{Set J-LIWC2015 dictionary path}
\usage{
setup_jliwcdic(dir = getOption("jliwc_dic_home"))
}
\arguments{
\item{dic}{path to the J-LIWC2015 dictionary}
}
\description{
Set J-LIWC2015 dictionary path
}
\examples{
setup_jliwcdic()

}
