% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_userdic.R
\name{setup_userdic}
\alias{setup_userdic}
\title{Setup J-LIWC2015 user dictionary}
\usage{
setup_userdic(
  dir = getOption("jliwc_project_home"),
  user_dic = "user_dict.dic",
  userdic_url = getOption("jliwc_USERDIC_url"),
  check = TRUE
)
}
\arguments{
\item{dir}{home directory for J-LIWC2015}

\item{user_dic}{filename of the user dictionary file}

\item{userdic_url}{URL to the user dictionary file}

\item{check}{check if the user dictionary file is properly installed}
}
\description{
Setup J-LIWC2015 user dictionary
}
\examples{
setup_userdic()

}
