% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize2.R
\name{tokenize2}
\alias{tokenize2}
\title{Function that calls \code{\link[gibasa:tokenize]{tokenize}}
with locally installed MeCab}
\usage{
tokenize2(x, ...)
}
\arguments{
\item{x}{Character vector}

\item{...}{Other arguments passed to \code{\link[gibasa:tokenize]{tokenize}}}
}
\value{
tokenized character vector
}
\description{
Function that calls \code{\link[gibasa:tokenize]{tokenize}}
with locally installed MeCab
}
