% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_dictionaries.R
\name{uninstall_dictionaries}
\alias{uninstall_dictionaries}
\title{Uninstall dictionary files
#'}
\usage{
uninstall_dictionaries(ipadic = TRUE, userdic = TRUE, jliwcdic = FALSE)
}
\arguments{
\item{ipadic}{A logical value to remove IPADic files}

\item{userdic}{A logical value to remove user dictionary files}

\item{jliwcdic}{A logical value to remove JLIWC dictionary files}
}
\value{
Boolean, \code{TRUE} if the setup is successful,
\code{FALSE} otherwise
}
\description{
Uninstall dictionary files installed by the jliwc package
}
\examples{
\dontrun{
uninstall_dictionaries()
}

}
