% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_info2.R
\name{dictionary_info2}
\alias{dictionary_info2}
\title{Function that calls \code{\link[gibasa:dictionary_info]{dictionary_info}}
with locally installed MeCab}
\usage{
dictionary_info2(...)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[gibasa:dictionary_info]{dictionary_info}}}
}
\value{
Directory information of the dictionary
}
\description{
Function that calls \code{\link[gibasa:dictionary_info]{dictionary_info}}
with locally installed MeCab
}
