% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liwc_analysis.R
\name{liwc_analysis}
\alias{liwc_analysis}
\title{LIWC-style analysis of Japanese text data}
\usage{
liwc_analysis(
  input,
  text_field = "text",
  dict = getOption("jliwc_dictfile"),
  lang = c("en", "ja"),
  pos_tag = TRUE,
  sys_dic = getOption("jliwc_IPADIC"),
  user_dic = getOption("jliwc_USERDIC")
)
}
\arguments{
\item{input}{A data frame or a list of character vector to analyze}

\item{text_field}{A column name of text data (if input is a data frame)}

\item{dict}{A dictionary2 class object of J-LIWC2015 dictionary}

\item{lang}{A character vector specifying language of the output tag (English or Japanese)}

\item{pos_tag}{logical; if TRUE, add POS percentage in MeCab}

\item{sys_dic}{A path to a MeCab system dictionary}

\item{user_dic}{A path to a MeCab user dictionary}
}
\value{
A data frame with LIWC-style analysis results. Rows are LIWC categories.
}
\description{
LIWC-style analysis of Japanese text data
}
\examples{
\dontrun{
setup_ipadic()
setup_userdic()
setup_jliwcdic()

# Sample Japanese texts from Night on the Galactic Railroad (Kenji Miyazawa)
x <- gibasa::ginga[1:10]
x

liwc_results <- x |> liwc_analysis()
liwc_results

# data frame
# The column name is "text" by default
# Each row is a unique text
texts_df <- data.frame(text = x)

liwc_results_df <- texts_df |> liwc_analysis()

dplyr::tibble(liwc_results_df) # Same output as liwc_results
}

}
