% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing text for morphological analysis}
\usage{
preprocess(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A character vector
}
\description{
Preprocessing text for morphological analysis
}
\examples{
x <- preprocess("\uff2f\uff2b\uff1f\uff12\u6642\u9593\u5f8c\u306b\uff01")
x

}
