% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dict.R
\name{read_dict}
\alias{read_dict}
\title{Read the LIWC dictionary}
\usage{
read_dict(file, format = c("LIWC2015", "LIWC22"))
}
\arguments{
\item{file}{A file path of the dictionary}

\item{format}{A format of the dictionary (LIWC2015 or LIWC22)}
}
\value{
A dictionary object
}
\description{
Read the LIWC dictionary
}
\examples{
\dontrun{
# Set up dictionaries
setup_ipadic()
setup_userdic()
setup_jliwcdic() # choose a dictionary file from a local directory

# Read the dictionary (LIWC2015 format)
dictliwc <- read_dict(file.path(getOption("jliwc_project_home"), "Japanese_Dictionary.dic"))
}

}
