% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_jliwcdic.R
\name{setup_jliwcdic}
\alias{setup_jliwcdic}
\title{Set J-LIWC2015 dictionary path}
\usage{
setup_jliwcdic(
  dir = getOption("jliwc_project_home"),
  format = getOption("jliwc_format", default = "LIWC2015")
)
}
\arguments{
\item{dir}{a path to the J-LIWC2015 dictionary}

\item{format}{A format of the dictionary (LIWC2015 or LIWC22)}
}
\value{
Boolean, \code{TRUE} if the setup is successful,
\code{FALSE} otherwise
}
\description{
Set J-LIWC2015 dictionary path
}
\examples{
\dontrun{
setup_jliwcdic()
}

}
