% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize2.R
\name{tokenize2}
\alias{tokenize2}
\title{function that always set the temporary mecabrc environment variable}
\usage{
tokenize2(x, ...)
}
\arguments{
\item{x}{character vector}

\item{...}{other arguments passed to \code{\link[gibasa]{tokenize}}}
}
\value{
tokenized character vector
}
\description{
function that always set the temporary mecabrc environment variable
}
